chore: organzing react query
