+++ apis/auth/index.ts
@@ -1,0 +1,1 @@
+export { login, logout } from './mutations';
